refactor(app): extract the palette and settings wiring
The command palette and every settings-modal toggle (theme, update
check, sidebar side, font size, history retention, NoSQL collection cap,
auto table alias, query-error highlight) move to `wire::settings`.

`rebuild_query_tree` gains an explicit `Rc<dyn Fn(&str)>` type so it can
be handed to a wiring module rather than only captured in place.
