+++ app/src/wire/mod.rs
@@ -5,4 +5,5 @@
 //! and calls them in order; the handler bodies themselves are unchanged.
 
 pub(crate) mod conn_form;
+pub(crate) mod settings;
 pub(crate) mod update;
